*** empty log message ***

+++ src/unixatomic.c
@@ -57,7 +57,7 @@ ssize_t atomic_read(int fd, const void *buf, size_t count)
 {
   char *b = (char *) buf;
   ssize_t bytes_read = 0;
-  int ret;
+  ssize_t ret;
   while (bytes_read < count) {
     ret = read(fd, &b[bytes_read], count - bytes_read);
     if (ret < 0) {
@@ -128,7 +128,7 @@ ssize_t atomic_write(int fd, const void *buf, size_t count)
 {
   char *b = (char *) buf;
   ssize_t bytes_written = 0;
-  int ret;
+  ssize_t ret;
   while (bytes_written < count) {
     ret = write(fd, &b[bytes_written], count - bytes_written);
     if (ret < 0) {
